Bugfix: avatar border shows with no avatar
If the site doesn't specify an avatar logo, the border will still show,
but the avatar itself will be empty, causing a weird looking ellipse
thing to be rendered on the page.

This encapsulates the whole of the avatar html in the check for the
site logo, causing none of it to be rendered when there isn't one.